Changed INT to SMALLINT where appropriate.

+++ src/Northwind/SqlServer/Tables/06. Territories.sql
@@ -2,6 +2,6 @@
 CREATE TABLE Territories (
     TerritoryId             NVARCHAR(20)    NOT NULL,
     TerritoryDescription    NCHAR(50)       NOT NULL,
-    RegionId                INT             NOT NULL,
+    RegionId                SMALLINT        NOT NULL,
     CONSTRAINT PK_Territories PRIMARY KEY NONCLUSTERED (TerritoryId ASC)
 );
